SpinBox: Only call fixup on content change in live mode
We shouldn't call fixup on intermediate or valid results while the user
is editing; at least not if the live proverty is false, as there's no
need to always have the latest value available.

Unconditionally calling fixup means that it is no longer possible to
have a temporarily invalid input. We could check whether the state is
Invalid (as opposed to Intermediate), and only call fixup then, but even
that is probably unnecessary overhead. Especially given that completely
invalid input (e.g. a letter when a number is expected) is still blocked
anyway, as this already happens via TextInput's validator.
The expectation is that a style properly shares the validator between the
SpinBox and its TextInput, like Qt's own style do.

The behavior stays the same as it was before when "live" is set to true.
The behavior in that case is basically as if one would press Enter
after every change.

Amends ce85cb16160394d616bc3dc4bf2a8915ec63b036

Fixes: QTBUG-129939
Change-Id: I583b472486ca3a5e8109d73a9e5c9947016aec98
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 61aae86b5f0f21041095c2b28376ec0cbbc83a15)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>