aotstats: Support --only-bytecode and modules with no qml files
It is possible to pass --only-bytecode to qmlcachegen. As the name
implies, this only generates the bytecode for the qml files and does not
compile them. This case was not taken into account which could lead to
files commands depend upon not being generated.

Therefore, keep track of empty and only-bytecode modules in files
generated by cmake and pass them to qmlaotstats upon aggregation such
that it can include that information in the report.

Also, only pass the arguments specific to aotstats to qmlcachegen if
--only-bytecode is not set for that module.

Fixes: QTBUG-130084
Task-number: QTBUG-124667
Change-Id: I44b4a80e8a6fd2f9bc16ae1bb2c8d540ff3b697b
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit 318ac440653a570d2bd18fc3d25e1b2fc8af88fa)
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>