Engine: Keep script context alive
A closure might need to access the imported scripts, which live in the
context. The closure might however live longer than than the context is
valid, and so far, the lookup would have failed.

Solve that issue by not completely discarding the reference to the
script, but instead downgrading it to a weak reference. Additionally,
closurse will keep a reference to the scripts as an invalid property, so
that the gc will keep them alive.

The ignoreMessage call in
importedScriptsAccessOnObjectWithInvalidContext is removed, as we keep
the script value alive.

Fixes: QTBUG-130575
Change-Id: Ieb3044c419271473a9fc9afcf9e88ab4b3f4b88a
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit aefdce82fd53537c7a2f4c3d63714d1e0d335551)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>